Refactor installation of Python requirements.
The current code installs the Python requirements files in a different order
than "paver install_prereqs".  This has been causing problems for the
edx-solutions fork of edx-platform.

The code is also rather repetitive.  This patch cleans the code up a bit, while
at the same time making sure that the order in which the requirements files are
processed matches the order paver uses.

Additionally, the patch extends the list of requirements files which are scanned
for Github URLs by two items that seemed to be missing.