Make S3 and SWIFT log syncing independent.
While COMMON_OBJECT_STORE_LOG_SYNC and related variables are still
supported, this patch offers a finer grained control over tracking log
syncing.

Sometimes you want to host edX on openstack (and therefore need to run
the openstack role), but you still want to sync tracking logs to S3, so
it does not make sense to tie the choice of log sync store to the
presence of either aws or openstack role in the playbook.