Add support for LMS_INTERNAL_ROOT_URL.
This adds ability to configure the LMS_INTERNAL_ROOT_URL setting via the
EDXAPP_LMS_INTERNAL_ROOT_URL ansible variable. The value of the new
variable is used instead of EDXAPP_LMS_ROOT_URL to construct enterprise
API URLs, which are used for internal communication. In some deployment
scenarios, it is not desirable for internal communication to be carried
out via the external URLs.

This change is backwards compatible since the new EDXAPP_LMS_INTERNAL_ROOT_URL
defaults to the value of EDXAPP_LMS_ROOT_URL.